requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
asyncio>=3.4.3
pydantic>=2.0.0
starlette>=0.27.0
psutil>=5.9.0 